+++ src/app/(auth)/landing/layout.tsx
