--- src/app/(auth)/sign-in/[[...sign-in]]/_page.tsx
@@ -1,5 +1,0 @@
-import { SignIn } from "@clerk/nextjs";
-
-export default function SignInPage() {
-  return <SignIn />;
-}
